Error handling on '-p' option parsing.
Note: this code is intentionally made simple in order to avoid bloat.
As an example, we don't check PID_MAX_LIMIT (at least for Linux) for a
reason. Maximum pid is configurable in Linux as
/proc/sys/kernel/pid_max , but is no larger than PID_MAX_LIMIT.
However, users won't distinguish between PIDs that are off the limit
and PIDs that are unused. So we just check _our_ Hashtable limit
(UINT_MAX) instead.