Graph meter coloring by items. (Rebased commit, part 2)
This patch colors the graph meter based on the items of the bar meter.
For example, if most of the CPU power is running on low-priority
processes, then blue will be shown on the graph columns. If most of CPU
is running on kernel things, then the columns will be red, and so on.

This commit contains a linear time algorithm O(GRAPH_HEIGHT + items)
for determining color.